requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
idna==3.4
numpy==1.23.4
Pillow==9.3.0
PyYAML==6.0
requests==2.28.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
